requirements.txt
torch>=1.10
torchvision
timm
numpy
pandas
scikit-learn
matplotlib
albumentations
tqdm
jupyter
fastapi
uvicorn
pillow
pyyaml
streamlit
requests
ultralytics
opencv-python
